Zod coercion and protect player form actions

+++ src/routes/[tripId]/players/+page.server.ts
@@ -1,7 +1,7 @@
 import { error, invalid } from '@sveltejs/kit';
-import { supabase } from '$lib/supabaseClient';
 import type { Actions, PageServerLoad } from './$types';
 import { z } from 'zod';
+import { getSupabase } from '@supabase/auth-helpers-sveltekit';
 
 export const load: PageServerLoad = async ({ parent }) => {
 	const { title } = await parent();
@@ -11,18 +11,22 @@ export const load: PageServerLoad = async ({ parent }) => {
 };
 
 export const actions: Actions = {
-	addPlayer: async ({ request }) => {
+	addPlayer: async (event) => {
+		const { request } = event;
+		const { session, supabaseClient } = await getSupabase(event);
+		if (!session) throw error(403, { message: 'Unauthorized' });
+
 		const data = Object.fromEntries(await request.formData());
 		const playerSchema = z.object({
-			tripId: z.preprocess((id) => parseInt(z.string().parse(id)), z.number()),
+			tripId: z.coerce.number(),
 			name: z.string(),
-			handicap: z.preprocess((h) => parseFloat(z.string().parse(h)), z.number())
+			handicap: z.coerce.number()
 		});
 
 		try {
 			const { tripId: trip_id, name, handicap } = playerSchema.parse(data);
 
-			const { error: pgError } = await supabase
+			const { error: pgError } = await supabaseClient
 				.from('trip_players')
 				.insert({ trip_id, name, handicap });
 
@@ -31,19 +35,23 @@ export const actions: Actions = {
 			return invalid(400, { message: `failed to parse player, ${error}` });
 		}
 	},
-	updatePlayer: async ({ request }) => {
+	updatePlayer: async (event) => {
+		const { request } = event;
+		const { session, supabaseClient } = await getSupabase(event);
+		if (!session) throw error(403, { message: 'Unauthorized' });
+
 		const data = Object.fromEntries(await request.formData());
 		const playerSchema = z.object({
-			tripId: z.preprocess((id) => parseInt(z.string().parse(id)), z.number()),
-			playerId: z.preprocess((id) => parseInt(z.string().parse(id)), z.number()),
+			tripId: z.coerce.number(),
+			playerId: z.coerce.number(),
 			name: z.string(),
-			handicap: z.preprocess((h) => parseFloat(z.string().parse(h)), z.number())
+			handicap: z.coerce.number()
 		});
 
 		try {
 			const { tripId, playerId, name, handicap } = playerSchema.parse(data);
 
-			const { error: pgError } = await supabase
+			const { error: pgError } = await supabaseClient
 				.from('trip_players')
 				.update({ name, handicap })
 				.eq('trip_id', tripId)
@@ -54,17 +62,21 @@ export const actions: Actions = {
 			return invalid(400, { message: `failed to parse player, ${error}` });
 		}
 	},
-	deletePlayer: async ({ request }) => {
+	deletePlayer: async (event) => {
+		const { request } = event;
+		const { session, supabaseClient } = await getSupabase(event);
+		if (!session) throw error(403, { message: 'Unauthorized' });
+
 		const data = Object.fromEntries(await request.formData());
 		const deleteSchema = z.object({
-			tripId: z.preprocess((id) => parseInt(z.string().parse(id)), z.number()),
-			playerId: z.preprocess((id) => parseInt(z.string().parse(id)), z.number())
+			tripId: z.coerce.number(),
+			playerId: z.coerce.number()
 		});
 
 		try {
 			const { tripId, playerId } = deleteSchema.parse(data);
 
-			const { error: pgError } = await supabase
+			const { error: pgError } = await supabaseClient
 				.from('trip_players')
 				.delete()
 				.eq('trip_id', tripId)
